move space config & info into space profile trait
