+++ contracts/space/Cargo.toml
@@ -27,6 +27,7 @@ std = [
     "ink/std",
     "scale/std",
     "scale-info/std",
+    "shared/std",
     "openbrush/std",
 ]
 ink-as-dependency = []
