+++ shared/src/traits/mod.rs
@@ -1,2 +1,3 @@
 pub mod codehash;
 pub mod plugin_launcher;
+pub mod space_profile;
