+++ shared/src/traits/plugin_launcher.rs
@@ -1,5 +1,5 @@
-pub use ink::prelude::string::String;
-pub use openbrush::{
+use ink::prelude::string::String;
+use openbrush::{
   modifiers,
   traits::{
     AccountId,
@@ -9,7 +9,7 @@ pub use openbrush::{
   storage::{Mapping},
   contracts::{ownable::OwnableError}
 };
-pub use crate::ensure;
+use crate::ensure;
 pub use crate::traits::plugin_launcher;
 
 pub type Version = u32;
